[♻️refactor]: HoverIcon을 Icons 폴더로 이동
