+++ src/components/KanbanBoard/CreateIssueBtn.tsx
@@ -2,7 +2,7 @@ import { css } from '@emotion/react';
 
 import { theme, ThemeType } from '@/styles/theme';
 
-import HoverIcon from '../Shared/HoverIcon';
+import HoverIcon from '../Shared/Icons/HoverIcon';
 import PlusIcon from '../Shared/Icons/PlusIcon';
 
 export default function CreateIssueBtn({ ...props }: React.ComponentProps<'button'>) {
