+++ src/components/gpt/info/APIKeyForm.tsx
@@ -4,7 +4,7 @@ import { toast } from 'react-toastify';
 import { useRecoilState } from 'recoil';
 
 import { gptAtom } from '@/atoms/gptAtom';
-import HoverIcon from '@/components/Shared/HoverIcon';
+import HoverIcon from '@/components/Shared/Icons/HoverIcon';
 import CheckIcon from '@/components/Shared/Icons/CheckIcon';
 import DeleteIcon from '@/components/Shared/Icons/DeleteIcon';
 import useInput from '@/hooks/useInput';
